Added UIKit and sandboxing of the modals content to the web loader
The sandboxing is done through the spawning of a new document. The content of every menu is an IFrame pointing to that document. When a Menu is opened the body of that document is replaced with the default content of the Menu, so that there is no need to reload UIKit. The goal of the sandboxing is to load UIKit and have a good looking interface without having to inject UIKit into the main page wich could break the styling of that page